Revert "Fix compiler warnings in ext/hash"
This reverts commit 596fbf7393777cb14006813111875db5b0a21d3b.

+++ ext/hash/hash_snefru.c
@@ -142,9 +142,9 @@ PHP_HASH_API void PHP_SNEFRUUpdate(PHP_SNEFRU_CTX *context, const unsigned char 
 	if ((MAX32 - context->count[1]) < (len * 8)) {
 		context->count[0]++;
 		context->count[1] = MAX32 - context->count[1];
-		context->count[1] = (uint32_t) (len * 8) - context->count[1];
+		context->count[1] = (len * 8) - context->count[1];
 	} else {
-		context->count[1] += (uint32_t) (len * 8);
+		context->count[1] += len * 8;
 	}
 
 	if (context->length + len < 32) {
